refactor(gui): extract Pod control + transport status into PodStatusMixin
Move the pod start/stop handlers, the FreeRDP launch path, and the 15s
polling timer that paints the pod-state chip + agent/RDP transport dots
off WinpodxWindow into a private mixin so main_window.py can focus on
overall window orchestration:

- PodStatusMixin (src/winpodx/gui/_main_window_pod.py)
  _launch_app (with _launch_lock class attr), _on_start_pod,
  _on_stop_pod, _start_status_timer, _refresh_pod_status,
  _on_transport_status, _on_pod_status, _on_app_launched,
  _on_app_launch_failed.

WinpodxWindow now reads
`class WinpodxWindow(AppCrudMixin, InfoPageMixin, LogsMixin, MaintenanceMixin, PodStatusMixin, QMainWindow)`.

The v0.2.0.10 auto-discovery-on-pod-up trigger and the launch-lock
serialization both come along as-is — they were behaviour bug fixes
encoded in the methods. No behavior change; tests: 1096 passed.

main_window.py: 1828 -> 1588 lines (-240). Cumulative: 2745 -> 1588 (-42%).

Remaining bulk is the _build_* UI builders (~1300 lines). A future
slice will split those by page rather than as one giant move, to
keep individual PRs reviewable.
